ciao-launcher: docker: Fix network deletion
The docker and VM network deletion code was mixed up.  This
meant we were trying to delete the docker network constructs when
deleting a VM and vice versa.  This didn't cause any problems
for VMs as the VNICs were still deleted but it meant that the
docker networks were leaked if you deleted a docker container.

This lead to start failures when trying to launch new containers
on the same node.  See the following issue for more details.

https://github.com/01org/ciao/issues/30

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>