ciao-controller: datastore: make error path for StartFailure faster
We can receive StartFailures while in the middle of issueing
many START commands.  As a result, it's important not to spend
too much time updating the usage and deleting the instances
during this path, because otherwise the cache is locked and new
instances cannot be added.  Previously we were a bit lazy
with deleting usage and just deleted it in the sql database
then got a fresh tenant usage by querying the database.  This
is way too slow.  this patch updates the usage in the cache
without requiring sql interaction.  We also put the sql udpate
for the persistent store in a separate go func so that it can
be done asynchronously out of the fast path.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>